+++ playwright.behavior.config.js
@@ -16,6 +16,8 @@ module.exports = defineConfig({
 		"link-visited.spec.js",
 		"list-nesting.spec.js",
 		"nav-dropdown.spec.js",
+		"prefers-contrast.spec.js",
+		"print.spec.js",
 	],
 	fullyParallel: true,
 	forbidOnly: !!process.env.CI,